+++ Modules/Intent.Modules.Common.CSharp/TypeRegistry.cs
@@ -1,105 +1,113 @@
-﻿using System.Collections.Generic;
+﻿#nullable enable
+using System.Collections.Generic;
 using System.Linq;
 
 namespace Intent.Modules.Common.CSharp;
 
 internal class TypeRegistry
 {
-    private readonly Member _members = new();
+    private readonly TypePathPart _globalNamespace = new(isType: false, parent: null);
 
-    public TypeRegistry(IEnumerable<string> paths)
+    public TypeRegistry WithTypes(IEnumerable<string> fullyQualifiedTypeNames)
     {
-        foreach (var path in paths)
+        foreach (var fullyQualifiedTypeName in fullyQualifiedTypeNames)
         {
-            var current = _members;
-
-            var split = path.Split('.');
-            for (var i = 0; i < split.Length; i++)
-            {
-                var item = split[i];
-                if (current.ContainsKey(item))
-                {
-                    current = current[item];
-                }
-                else
-                {
-                    var member = new Member(i == split.Length - 1);
-                    current[item] = member;
-                    current = member;
-                }
-            }
+            Add(fullyQualifiedTypeName);
         }
+
+        return this;
+    }
+
+    public TypeRegistry WithNamespaces(IEnumerable<string> namespaces)
+    {
+        foreach (var @namespace in namespaces)
+        {
+            Add(@namespace, typeName: null);
+        }
+
+        return this;
     }
 
     public void Add(string fullyQualifiedTypeName)
     {
-        var lastIndexOf = fullyQualifiedTypeName.LastIndexOf('.');
-        Add(fullyQualifiedTypeName[..lastIndexOf], fullyQualifiedTypeName[(lastIndexOf + 1)..]);
+        var split = fullyQualifiedTypeName.Split('.');
+        Add(split[..^1], split[^1..]);
     }
 
-    public void Add(string @namespace, string typeName)
+    public void Add(string @namespace, string? typeName)
     {
-        var current = _members;
+        Add(@namespace.Split('.'), typeName == null ? [] : [typeName]);
+    }
 
-        var split = @namespace.Split('.');
-        foreach (var item in split)
+    public void Add(IEnumerable<string> namespaceParts, IEnumerable<string> typeNameParts)
+    {
+        var current = _globalNamespace;
+
+        foreach (var part in namespaceParts)
         {
-            if (current.ContainsKey(item))
+            if (current.ContainsKey(part))
             {
-                current = current[item];
+                current = current[part];
             }
             else
             {
-                current = new Member();
-                current[item] = current;
+                var typePartPath = new TypePathPart(isType: false, parent: current);
+                current[part] = typePartPath;
+                current = typePartPath;
             }
         }
-        current[typeName] = new Member(true);
-    }
 
-    public bool Contains(string @namespace)
-    {
-        var current = _members;
-
-        var split = @namespace.Split('.');
-        return split.All(item => current.TryGetValue(item, out current));
-    }
-
-    public bool ContainsType(string typeName)
-    {
-        var current = _members;
-
-        var split = typeName.Split('.');
-        foreach (var item in split)
+        foreach (var part in typeNameParts)
         {
-            if (!current.TryGetValue(item, out current))
+            if (current.ContainsKey(part))
             {
+                current = current[part];
+                current.IsType = true;
+            }
+            else
+            {
+                var typePartPath = new TypePathPart(isType: true, parent: current);
+                current[part] = typePartPath;
+                current = typePartPath;
+            }
+        }
+    }
+
+    public bool Contains(string path, out bool isType)
+    {
+        return Contains(path.Split('.'), out isType);
+    }
+
+    public bool Contains(string @namespace, string typeName, out bool isType)
+    {
+        return Contains(@namespace.Split('.').Append(typeName), out isType);
+    }
+
+    private bool Contains(IEnumerable<string> typeParts, out bool isType)
+    {
+        var current = _globalNamespace;
+
+        foreach (var part in typeParts)
+        {
+            if (!current.TryGetValue(part, out current))
+            {
+                isType = default;
                 return false;
             }
         }
 
-        return current.IsType;
+        isType = current.IsType;
+        return true;
     }
 
-
-    public bool Contains(string @namespace, string typeUnqualified)
+    private class TypePathPart(bool isType, TypePathPart? parent) : Dictionary<string, TypePathPart>
     {
-        var current = _members;
+        private bool _isType = isType;
 
-        var split = @namespace.Split('.');
-        foreach (var item in split.Append(typeUnqualified))
+        public bool IsType
         {
-            if (!current.TryGetValue(item, out current))
-            {
-                return false;
-            }
+            get => _isType || parent?.IsType == true;
+            set => _isType = value;
         }
-
-        return current.Count == 0;
-    }
-
-    private class Member(bool isType = false) : Dictionary<string, Member>
-    {
-        public bool IsType { get; } = isType;
     }
 }