+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,5 +1,6 @@
 ﻿### Version 3.8.2
 
+- Fixed: Required type disambiguation would not occur when the conflicting type was in the current full namespace.
 - Fixed: Return types of IAsyncEnumerable are now not automatically wrapped in Task/ValueTask when method is marked as async
 
 ### Version 3.8.1